lots of fixes, cleanup, minor improvements
